Fixed Merge
Values of merged df in the origen dataset was not correct. When doing a search on 'Simon Sinek', for example, we obtained 1 talk from him and 2 that were not (ntroducing Body Stuff with Dr. Jen Gunter & Save the oceans, feed the world!). This error was because the merge was done on the left side with the larger dataset as the left object, as this:

df=df_scrapped.merge(df_kaggle, left_on='link', right_on='link')

 We modified this by basing the merge on the scrapped dataset that is poised to be smaller. In this, only one talk from Simon Sinek is merged properly and this should be validated across all analysis.

Future enhancements would be to review the error log occurred during the merge ans try to rescue as many as possible